Fix Descent 2 delayed crash when movies are enabled
Commit 479884288c21 unintentionally caused the movie window to linger in
the background after a movie was played, which causes a crash when the
window is used later, because part of its state has been cleared.
Use window::track to detect its termination in the usual way, and
eliminate the special logic that tried to track the window by deferring
its deletion.  This causes the window to close at the proper time,
thereby avoiding a later crash.

Fixes: 479884288c21841fc04036635ccac2678a47604c ("Always delegate to window_create for send_creation_events")