Fix high score background truncation
Commit 4d6affd74d33 ("Fix SDL-only briefing crash") added a hack to
prevent a crash when the briefing window is automatically resized.
However, a logic error in that hack causes some subcanvases that were
not wrong to be truncated.  Change the logic to fix those subcanvases,
such as the one used for rendering a background in the High Scores
window.  The new logic still avoids the crash in in SDL-only mode.

Fixes: 4d6affd74d33df6ee2c081b56159caa245c7fa10 ("Fix SDL-only briefing crash")