Lift thief AIM_THIEF_ATTACK check into caller
attempt_to_steal_item_3 always fails if the thief is not in
AIM_THIEF_ATTACK mode.  The caller will call attempt_to_steal_item_3
repeatedly in some cases, but the mode does not change during the calls,
so if one fails for this reason, then all will fail for this reason.
Lift the check into the caller so that the loop of repeated calls can be
skipped if the mode check fails.