Remove abandoned Visual Studio 2017 support
This never worked, but was committed to git in the hope it would be
useful for someone who wanted to finish it.  In the 5.5 years since
then, the closest to that has been one user who requested Visual Studio
2019 support[1], but went silent after reporting that unspecified errors
occurred.  The error messages were never shared.

Remove these project files, since they are not useful, and have caused
at least one report where the reporter incorrectly attached the stub
Visual Studio 2017 dxxsconf.h, instead of attaching the dxxsconf.h
generated by the build for which the reporter wanted assistance.

[1]: https://github.com/dxx-rebirth/dxx-rebirth/issues/607