Make sure gun point is calculated when firing
Gun point calculation is skipped in do_ai_frame if
dist_to_player >= 200 but a bot might still fire a homing weapon
if it is further away.
Add extra call to calc_gun_point in ai_do_actual_firing_stuff when
a homing weapon is fired without visibility to make sure a valid
gun point is used.

This is an original game bug, shown for example in this video:
https://www.youtube.com/watch?v=2dP51znGh7o