Add compile-time sanity check on multi_send_data_direct length
JoeNotCharles reported
<https://github.com/dxx-rebirth/dxx-rebirth/pull/693/commits/1f1903a3b98000719c200454fdc87ef52741eec4>
an overflow in `net_udp_send_mdata_direct`.  This overflow is impossible
as currently written, because it can occur only if
`multi_send_data_direct` passes an oversized buffer to
`net_udp_send_mdata_direct`, and no messages are large enough to trigger
this.  JoeNotCharles proposed adding a runtime check to abort the
program if this happens.  Instead, this commit adds a compile-time check
to detect use of an excessively large input buffer.