Simplify test for support of explicitly deleted functions
Support for explicitly deleted functions has been mandatory for a long
time.  Historically, this was handled through a special SConf test to
work around a clang-3.4 bug that warned when a deleted function had
named parameters.  Support for clang-3.4 was dropped some time ago, and
newer versions do not require this workaround.  Simplify the SConf
script by declaring explicitly deleted functions a required feature, so
that they can be tested as a group with the other required features.