Reduce list copying in environment overrides
Prefer creating a copy explicitly, then appending or extending it as
needed, rather than using `__add__` and allowing it to duplicate the
sequence.  `__add__` has two undesirable properties in this script:

- Using `A + B + C` can create multiple temporaries unnecessarily.
- `A + B` typically requires that `A` and `B` be the same type of
  sequence, but this script does not need to require that.  Using an
  explicit copy+extend allows `A` and `B` to be different types of
  sequence.

Make use of the latter property to turn some member variables that were
`list` for the sake of `__add__`'s type requirement into `tuple`, since
they are never rewritten.